+++ src/ArgentPonyWarcraftClient/Models/GameDataApi/AuctionHouse/AuctionItem.cs
@@ -11,19 +11,19 @@ namespace ArgentPonyWarcraftClient
         /// Gets the ID of the auction item.
         /// </summary>
         [JsonPropertyName("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         /// <summary>
         /// Gets the contex for the item.
         /// </summary>
         [JsonPropertyName("context")]
-        public long? Context { get; set; }
+        public int? Context { get; set; }
 
         /// <summary>
         /// Gets the bonus lists for the item.
         /// </summary>
         [JsonPropertyName("bonus_lists")]
-        public long[] BonusLists { get; set; }
+        public int[] BonusLists { get; set; }
 
         /// <summary>
         /// Gets the modifiers for the item.
@@ -35,24 +35,24 @@ namespace ArgentPonyWarcraftClient
         /// Gets the ID of the pet breed, if the item is a pet.
         /// </summary>
         [JsonPropertyName("pet_breed_id")]
-        public long? PetBreedId { get; set; }
+        public int? PetBreedId { get; set; }
 
         /// <summary>
         /// Gets the pet level, if the item is a pet.
         /// </summary>
         [JsonPropertyName("pet_level")]
-        public long? PetLevel { get; set; }
+        public int? PetLevel { get; set; }
 
         /// <summary>
         /// Gets the ID of the pet quality ID, if the item is a pet.
         /// </summary>
         [JsonPropertyName("pet_quality_id")]
-        public long? PetQualityId { get; set; }
+        public int? PetQualityId { get; set; }
 
         /// <summary>
         /// Gets the ID of the pet species ID, if the item is a pet.
         /// </summary>
         [JsonPropertyName("pet_species_id")]
-        public long? PetSpeciesId { get; set; }
+        public int? PetSpeciesId { get; set; }
     }
 }
